cli/requirements.txt
requests
pycryptodome
pyotp
